+++ packages/@expo/cli/src/export/exportApp.ts
@@ -317,6 +317,8 @@ export async function exportAppAsync(
           platform: 'web',
           apiRoutesOnly: !isWeb,
           templateHtml,
+          outputDir: outputPath,
+          baseUrl,
         });
       }
 
@@ -371,6 +373,8 @@ export async function exportAppAsync(
           files,
           platform: 'web',
           apiRoutesOnly: true,
+          outputDir: outputPath,
+          baseUrl,
         });
 
         // Output a placeholder index.html if one doesn't exist in the public directory.